pyproject.toml
[project]
name = "mimmy"
version = "0.0.1"
description = "Single-ticker focused LLM trading agent team"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "anthropic>=0.40",
    "openai>=1.40",
    "google-genai>=0.3",
    "python-telegram-bot[ext]>=21.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "structlog>=24.1",
    "sqlmodel>=0.0.16",
    "boto3>=1.34",
    "tenacity>=8.2",
    "typer>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
mimmy = "mimmy.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mimmy"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
